+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - Removed `protocol::digest` re-export ([#75]).
+- `digest` and `signature` are now re-exported from the top level instead of `session` ([#75]).
 
 
 [#75]: https://github.com/entropyxyz/manul/pull/75